requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
openai
assemblyai
pytube
yt-dlp
